Experimenting with a new data model to store query stats
As of now each query entry stores client-ip, bucket-id application_name_hash,
user_id, database_id along with query hash. That effectively means we
store 40 bytes of meta data for each distinct query. Moreover, the distinct
query means the combination of bucket-id, client-ip, application-name-hash,
user-id and database-id is unique. Now as we are trying to support
IPV6 client addresses as well in future that would take up the metadata size
to roughly 80 bytes for each query entry, which is quite a lot.

This commit tries to tackle this issue by taking the connection related metadata
storage out-of-line in separate hashtable.  One big assumption is that in the
majority of cases the 'number of queries' > 'number of distinct
user-database-client-ip' combinations, secondly the client-ip and database-id
remains constant for the lifetime of the connection while user-id and application
name rarely changes. So on the basis of these two assumptions the hash table
that stores the connection related information (We call it Origines in the code)
uses a separate lock and offloads that processing from the pgss_store() function
and performs it once in the post parse analyze stage.In terms of data saved for
each query entry, only BucketID and OriginID, along with QueryID gets stored in
each query entry while OriginInfo entry is shared among multiple query entries.

I am also expecting some good performance gains along with memory usage
improvements in high load environments with considerable concurrent connections
with this change but that still requires some more benchmarking.

Finally, the regression after the commit is fine except for the 
application_name_unique test case, which IMHO is now producing more accurate output.